# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "baas",
    "version": "0.2.0"
  },
  "paths": {
...
    "/platform/accounts/{accountReference}/transactions": {
      "get": {
        "operationId": "operatorListTransactions",
        "summary": "Any account's transactions. Audited.",
        "responses": {
          "200": {
            "description": "A page of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPage"
                }
              }
            }
          },
          "404": {
            "description": "No such account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/platform/api-clients/{id}/scopes": {
      "get": {
        "operationId": "scopeHistory",
        "summary": "Every grant this client has ever had. Revoked grants are kept.",
        "responses": {
          "200": {
            "description": "The grant history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeHistory"
                }
              }
            }
          },
          "404": {
            "description": "No such API client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "grantScope",
        "summary": "Grant one scope, with a reason. There is deliberately no route that takes a list.",
        "responses": {
          "201": {
            "description": "Granted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantedScope"
                }
              }
            }
          },
          "400": {
            "description": "Unknown scope, or no reason given",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such API client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "That scope is already granted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platform/api-clients/{id}/scopes/{scope}": {
      "delete": {
        "operationId": "revokeScope",
        "summary": "Revoke one scope. The grant row is stamped, never deleted.",
        "responses": {
          "200": {
            "description": "Whether a live grant was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedScope"
                }
              }
            }
          },
          "404": {
            "description": "No such API client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platform/customers": {
      "get": {
        "operationId": "findCustomer",
        "summary": "Find one customer by an exact identifier. No listing, no prefix search.",
        "responses": {
          "200": {
            "description": "The customer and their provider links",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSummary"
                }
              }
            }
          },
          "400": {
            "description": "Supply exactly one identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/platform/provider-requests/{id}": {
      "get": {
        "operationId": "getProviderRequest",
        "summary": "One provider call, including the scrubbed request and response bodies. Separately audited.",
        "responses": {
          "200": {
            "description": "The provider call",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderCall"
                }
              }
            }
          },
          "404": {
            "description": "No such provider request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerSummary": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "externalUserUuid": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "providers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "providerId": {
                  "type": "string",
                  "maxLength": 64,
                  "pattern": "^[a-z][a-z0-9]*(?:[_-][a-z0-9]+)*$"
                },
                "status": {
                  "type": "string",
                  "minLength": 1
                },
                "statusReason": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "observedAt": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                }
              },
              "required": [
                "providerId",
                "status",
                "statusReason",
                "observedAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "customerId",
          "externalUserUuid",
          "providers"
        ],
        "additionalProperties": false
      },
      "ErrorResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9]*(?:[._][a-z0-9]+)*$"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "minLength": 1
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "code",
          "message",
          "correlationId"
        ],
        "additionalProperties": false
      },
      "GrantedScope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "granted": {
            "type": "string",
            "enum": [
              "mobile:accounts",
              "mobile:transactions"
            ]
          }
        },
        "required": [
          "granted"
        ],
        "additionalProperties": false
      },
...
      "ProviderCall": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "providerId": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z][a-z0-9]*(?:[_-][a-z0-9]+)*$"
          },
          "operation": {
            "type": "string",
            "minLength": 1
          },
          "correlationId": {
            "type": [
              "string",
              "null"
            ]
          },
          "idempotencyId": {
            "type": [
              "string",
              "null"
            ]
          },
          "outcome": {
            "type": "string",
            "enum": [
              "ok",
              "rejected",
              "unreachable"
            ]
          },
          "responseStatus": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "startedAt": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "durationMs": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "requestBody": {
            "type": "string"
          },
          "responseBody": {
            "type": "string"
          },
          "errorMessage": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "providerId",
          "operation",
          "correlationId",
          "idempotencyId",
          "outcome",
          "responseStatus",
          "startedAt",
          "durationMs",
          "requestBody",
          "responseBody",
          "errorMessage"
        ],
        "additionalProperties": false
      },
...
      "RevokedScope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "revoked": {
            "type": "boolean"
          }
        },
        "required": [
          "revoked"
        ],
        "additionalProperties": false
      },
      "ScopeHistory": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "grants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                },
                "scope": {
                  "type": "string",
                  "minLength": 1
                },
                "grantedAt": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                },
                "grantedBy": {
                  "anyOf": [
                    {
                      "type": "string",
                      "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "revokedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "revokedBy": {
                  "anyOf": [
                    {
                      "type": "string",
                      "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "reason": {
                  "type": "string",
                  "minLength": 1
                },
                "live": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "scope",
                "grantedAt",
                "grantedBy",
                "revokedAt",
                "revokedBy",
                "reason",
                "live"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "grants"
        ],
        "additionalProperties": false
      },
...
      "TransactionPage": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "transactionReference": {
                  "type": "string",
                  "minLength": 1
                },
                "accountReference": {
                  "type": "string",
                  "minLength": 1
                },
                "direction": {
                  "type": "string",
                  "enum": [
                    "debit",
                    "credit"
                  ]
                },
                "amount": {
                  "type": "object",
                  "properties": {
                    "amount": {
                      "type": "string",
                      "pattern": "^-?\\d+(?:\\.\\d+)?$"
                    },
                    "currency": {
                      "type": "string",
                      "enum": [
                        "AED",
                        "BHD",
                        "EGP",
                        "EUR",
                        "GBP",
                        "INR",
                        "JOD",
                        "JPY",
                        "KWD",
                        "OMR",
                        "PHP",
                        "PKR",
                        "SAR",
                        "TND",
                        "USD"
                      ]
                    }
                  },
                  "required": [
                    "amount",
                    "currency"
                  ],
                  "additionalProperties": false
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "pending",
                    "settled",
                    "rejected",
                    "reversed",
                    "unknown"
                  ]
                },
                "counterpartyName": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "narrative": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "occurredAt": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                }
              },
              "required": [
                "transactionReference",
                "accountReference",
                "direction",
                "amount",
                "status",
                "counterpartyName",
                "narrative",
                "occurredAt"
              ],
              "additionalProperties": false
            }
          },
          "nextCursor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
    }
  }
}